pyproject.toml
[project]
name = "best-football"
version = "0.1.0"
description = "Youth and high school football information platform"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28",
    "jinja2>=3.1",
    "fastapi>=0.115",
    "uvicorn>=0.34",
    "python-multipart>=0.0.20",
    "beautifulsoup4>=4.14.3",
    "pypdf>=6.6.2",
    "markdown>=3.10.1",
    "pyyaml>=6.0.3",
]

[project.optional-dependencies]
dev = ["pytest", "ruff"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
